Tasks, Goals and Journals are now associated with a User
Only logged in users can view tasks, goals and journals. Users can only view tasks, goals and journals created by them. If trying to view those pages while not logged in, you will be redirected to the log in page. Remember to re-run the migrations for the changes to work.