To Do List improvements - Goal Association
Tasks are now organized by their corresponding goal on the to do list page. 

Streaks are displayed on the to do list page for corresponding goals. 

Tasks added from the to do list page must now be associated with a goal; a dropdown list of all goals is available when creating a new task here.